requirements.txt
requests==2.32.3
gTTS==2.5.3
trafilatura==2.0.0
beautifulsoup4==4.12.3
python-dotenv==1.0.1
deepl==1.19.1
moviepy==1.0.3
Pillow>=11.1.0
google-api-python-client==2.151.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
tweepy>=4.15.0
yfinance==0.2.54
